+++ docs/generate_docs.sh
@@ -2,8 +2,11 @@
 
 doxygen doxygen/Doxyfile  # doxygen docs
 
-epydoc --config epydoc/Epydoc_dvi  # epydoc docs
-epydoc --config epydoc/Epydoc_html
+epydoc --config epydoc/Epydoc_html  # epydoc docs
 epydoc --config epydoc/Epydoc_latex
 epydoc --config epydoc/Epydoc_pdf
-epydoc --config epydoc/Epydoc_ps
+
+cd sphinx  # sphunx docs
+make html
+make latex
+cd _build/latex && make  # compile latex docs